# src/main/java/com/example/exchangeapp/dto/ExchangeRateDto.java
package com.example.exchangeapp.dto;

import com.example.exchangeapp.enums.Currency;

public class ExchangeRateDto {
    private Currency from;
    private Currency to;
    private double buyRate;
    private double saleRate;

    public Currency getFrom() {
        return from;
    }

    public void setFrom(Currency from) {
        this.from = from;
    }

    public Currency getTo() {
        return to;
    }

    public void setTo(Currency to) {
        this.to = to;
    }

    public double getBuyRate() {
        return buyRate;
    }

    public void setBuyRate(double buyRate) {
        this.buyRate = buyRate;
    }

    public double getSaleRate() {
        return saleRate;
    }

    public void setSaleRate(double saleRate) {
        this.saleRate = saleRate;
    }
}