# src/main/java/com/example/exchangeapp/dto/ExchangeRequestDto.java
package com.example.exchangeapp.dto;

import com.example.exchangeapp.enums.Currency;
import com.example.exchangeapp.enums.OperationType;

public class ExchangeRequestDto {
    private double amountFrom;
    private double amountTo;
    private Currency currencyFrom;
    private Currency currencyTo;
    private OperationType operationType;

    public double getAmountFrom() {
        return amountFrom;
    }

    public void setAmountFrom(double amountFrom) {
        this.amountFrom = amountFrom;
    }

    public double getAmountTo() {
        return amountTo;
    }

    public void setAmountTo(double amountTo) {
        this.amountTo = amountTo;
    }

    public Currency getCurrencyFrom() {
        return currencyFrom;
    }

    public void setCurrencyFrom(Currency currencyFrom) {
        this.currencyFrom = currencyFrom;
    }

    public Currency getCurrencyTo() {
        return currencyTo;
    }

    public void setCurrencyTo(Currency currencyTo) {
        this.currencyTo = currencyTo;
    }

    public OperationType getOperationType() {
        return operationType;
    }

    public void setOperationType(OperationType operationType) {
        this.operationType = operationType;
    }
}