# src/main/java/com/example/exchangeapp/models/CommissionModel.java
package com.example.exchangeapp.models;

import com.example.exchangeapp.enums.Currency;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;

@Entity(name = "commission")
public class CommissionModel {
    @Id
    @GeneratedValue(strategy= GenerationType.IDENTITY)
    private Integer id;
    @Column(name = "rate")
    private double commission;
    @Column(name = "commission_from")
    private Currency from;
    @Column(name = "commission_to")
    private Currency to;

    public Integer getId() {
        return id;
    }

    public double getCommission() {
        return commission;
    }

    public void setCommission(double commission) {
        this.commission = commission;
    }

    public Currency getFrom() {
        return from;
    }

    public void setFrom(Currency from) {
        this.from = from;
    }

    public Currency getTo() {
        return to;
    }

    public void setTo(Currency to) {
        this.to = to;
    }
}